+++ doc/conf.py
@@ -33,11 +33,18 @@ sys.path.insert(0, os.path.abspath('../'))
 # ones.
 extensions = [
     'sphinx.ext.autodoc',
+    'sphinx.ext.intersphinx',
     'sphinx.ext.mathjax',
     'sphinx.ext.napoleon',
     'sphinx.ext.viewcode',
     'matplotlib.sphinxext.plot_directive'
 ]
+
+intersphinx_mapping = {
+    'numpy': ('https://numpy.org/doc/stable', None),
+    'python': ('https://docs.python.org/3', None),
+    'scipy': ('https://docs.scipy.org/doc/scipy', None),
+}
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ['_templates']